resolve all sign-compare warning

+++ src/common/file/file_path.cc
@@ -80,7 +80,7 @@ bool CreateDirWithRetry(const std::string& dir_path) {
     if (dir_path[0] == '/') {
         path.append("/");
     }
-    for (int d = 0; d < path_sections.size() && is_success; ++d) {
+    for (size_t d = 0; d < path_sections.size() && is_success; ++d) {
         if (path_sections[d] == ".") {
             continue;
         }
